+++ packages/language-tools/language-server/test/content-intellisense/caching.test.ts
@@ -5,7 +5,7 @@ import { after, before, describe, it } from 'node:test';
 import { Position } from '@volar/language-server';
 import { URI } from 'vscode-uri';
 import { getLanguageServer, type LanguageServer } from '../server.ts';
-import { fixtureDir } from '../utils.ts';
+import { fixtureDir } from '../test-utils.ts';
 
 const contentSchemaPath = path.resolve(fixtureDir, '.astro', 'collections', 'caching.schema.json');
