+++ packages/language-tools/language-server/test/misc/prettier-format.test.ts
@@ -3,7 +3,7 @@ import * as path from 'node:path';
 import { before, describe, it } from 'node:test';
 import { Range } from '@volar/language-server';
 import { getLanguageServer, type LanguageServer } from '../server.ts';
-import { fixtureDir } from '../utils.ts';
+import { fixtureDir } from '../test-utils.ts';
 
 describe('Formatting - Prettier', () => {
 	let languageServer: LanguageServer;
