+++ packages/language-tools/language-server/test/typescript/caching.test.ts
@@ -9,7 +9,7 @@ import {
 } from '@volar/language-server';
 import { URI } from 'vscode-uri';
 import { getLanguageServer, type LanguageServer } from '../server.ts';
-import { fixtureDir } from '../utils.ts';
+import { fixtureDir } from '../test-utils.ts';
 
 describe('TypeScript - Cache invalidation', async () => {
 	let languageServer: LanguageServer;
