+++ packages/language-tools/language-server/test/typescript/completions.test.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 import { before, describe, it } from 'node:test';
 import { Position } from '@volar/language-server';
 import { getLanguageServer, type LanguageServer } from '../server.ts';
-import { fixtureDir } from '../utils.ts';
+import { fixtureDir } from '../test-utils.ts';
 
 describe('TypeScript - Completions', async () => {
 	let languageServer: LanguageServer;
