+++ packages/language-tools/language-server/test/typescript/diagnostics.test.ts
@@ -4,7 +4,7 @@ import { before, describe, it } from 'node:test';
 import type { FullDocumentDiagnosticReport } from '@volar/language-server';
 import { type Diagnostic, DiagnosticSeverity, Range } from '@volar/language-server';
 import { getLanguageServer, type LanguageServer } from '../server.ts';
-import { fixtureDir } from '../utils.ts';
+import { fixtureDir } from '../test-utils.ts';
 
 describe('TypeScript - Diagnostics', async () => {
 	let languageServer: LanguageServer;
